feat(net): refactor gossip driver
